requirements.txt
pyyaml
torch==2.1.0
torchaudio==2.1.0
torchvision==0.16.0
ujson
ray==1.13.0
git+https://github.com/ildoonet/pytorch-gradual-warmup-lr.git
